+++ miso-web/src/it/resources/db/migration/clear_test_data.sql
@@ -14,6 +14,7 @@ DELETE FROM Transfer_Sample;
 DELETE FROM Transfer_Library;
 DELETE FROM Transfer_LibraryAliquot;
 DELETE FROM Transfer_Pool;
+DELETE FROM TransferChangeLog;
 DELETE FROM Transfer;
 DELETE FROM Workset_Sample;
 DELETE FROM Workset_Library;