+++ RELEASE_NOTES.md
@@ -8,7 +8,10 @@ Changes:
     * Library kit
     * Run sequencing kit
     * Container clustering kit
-    * Container multiplexing ki
+  * Log positions that box items are removed from
+  * Added change logs to transfers
+  * Fixed error that caused libraries to be missing from some sample contexts (including Children button and edit Sample page)
+  * Fixed error in LDAP configuration
   * Fixed creating samples parented to existing tissues with null times received and tube number
     (detailed sample)